4199 FIX Fix detection of replication mode if EC is distributed
If you use the new feature of having a distributed Event Console (consisting
of more than one Check_MK site) then the WATO display of the EC status
was broken. This has been fixed. WATO still shows only the status of
the local EC. This will be fixed in future.

Change-Id: Ife7d833e7ddcf65c8b1950a13c7235fe3e7a7a6e