4430 FIX Fixed logwatch-related Livestatus queries in NEB module
The 'mk_logwatch_files' and 'mk_logwatch_file' columns in the 'hosts' table
didn't work when the Nagios core was used, they were always empty.  The net
result was that logwatch-related things in the GUI always looked OK, even
when there were problems. This has been fixed.

Note that the CMC was not affected by this bug.

Change-Id: I9ebc41b848a857dd2513e488e3e323c862b2357f