4508 FIX Clusters: Now always have their nodes as parents (Fixes Nagios inconsistency)
When configuring clusters via Check_MK it was possible to configure custom parent nodes
for the cluster nodes when using the Nagios core. The Microcore already used it's nodes
before as it was designed. We have now changed the Nagios handling to also enforce the
nodes to be parent hosts of their cluster hosts.

The assumption is: The cluster host is reachable while at least one node is up.

Change-Id: Icb7c245a169129f1b55a2757bbe778d4ac882313