Made several row accesses more robust.
They worked only by accident, because some structs start with a field, so a
pointer to the struct and a pointer to the field have the same value.  This
is highly fragile, so we access the first field in type-safe manner now.

While we're there, use "auto" a bit more often, following Scott Meyer's
advice in "Effective Modern C++", item 5: "Prefer auto to explicit type
declarations."

Change-Id: I2dd8358046fadabd6070947db79073d5579ebf2e