4895 FIX cmk --convert-rrds: Don't fail when RRDs refered by PNP XMLs are missing
When using the "cmk --convert-rrds" command it is possible that XML files of PNP4Nagios
are refering to RRDs that do not exist. Previously the conversion totally failed with an
error message. It was now converted to a warning message instead. The conversion of other
RRDs is now continued in case of such errors.

Change-Id: I813b62ef9527646e3c67fd304e1a961105503080