4695 FIX Make communication with event console more robust
When the monitoring core connects to the event console very rapidly, a "No
buffer space available" error could happen intermittently.  This can be seen
in the log, and the corresponding command fails, too, e.g. archiving an
event.  This has been fixed, the connection is simply retried when this
rare but harmless error happens.

Change-Id: Id382dfe9f4267bb7d660ac306a963a3e622a8041