4914 FIX statgrab_mem: Plugin is now prioritized over solaris_mem if both is available
Until now the solaris_mem has been prioritized over statgrab_mem if both are available and deliver the same information. This has been changed by this werk. Since the memory info from top is under suspicion to be buggy we do prefer the mem information from statgrab from now on. The information from top will be a fallback only if statgrab is not available.

Change-Id: I120673e1c376e22ec8e90d75580a32e6a207d94d