5069 FIX Choose correct owner for report schedule entry when creating a new one
When you create a new entry for the report scheduler and you are in the
<tt>admin</tt> role then you can select the owner of this entry. But the
dialog would default to the user name being the first in the alphabet rather
then to your own name. This led to confusion and has now been fixed.

The problem showed up after adding an <tt>automation</tt> user per default,
which starts with the character <tt>a</tt> and this is most times the
first in the list.

Change-Id: I0b450e457a0389030043e2659bfeaf9d911af6d9