5051 FIX Fixed backup/restore of encrypted backups
It was not possible to perform encrypted backups since version 1.4.0b5. An error
message "NotImplementedError("Use module Crypto.Cipher.PKCS1_OAEP instead“)" was
displayed in the log when trying to create such a backup.

Workaround: Disable encrypted backups and perform unencrypted ones.

Encrypted backups created with previous versions need to be restored with 1.4.0b4
or older.

Change-Id: Ib3bd6798cc104b0ee1b1a5e9a7872223df7fe78b