5116 FIX if.include: Fixed wrong handling of duplicate interfaces if "use alias" or "use desccription" is configured
Within the rule {{Network Interface and Switch Port Discovery}}
you are able to configure the service description to use the
interface alias or description. Per default the index is used.
If there are duplicate interfaces with same name, the interface
index should be added.

Unfortunately, the afore-mentioned mechanism was broken. This
has been fixed.

Change-Id: I44221ac4b1323e551ce9758041fae192baa38405