5378 FIX Make Windows section webservices backwards compatible
Version 1.4.0 'corrected' the naming of section wmi_webservices.
Unfortunately, this broke the backwards compatibility of old ini files
that still call this 'webservices'. The fix maps the old name to the
new one so that both old and new ini files work. Moreover, the agent
bakery was still writing ini files with the old name. Now the agent
bakery is also fixed.

Change-Id: Iecfe16c55a6fdfb0b47b1a939a30d11c4707bc92