5261 FIX Improve agent bakery robustness when baking new agents
Depending on the complexity of the underlying configuration ruleset, some agents accidently could be marked as to be rebaked although there is no config change for them.
This is now mitigated by applying a more robust hashing algorithm for the agent config hashes.

After updating, no further action is required to apply the fix. Notably, no re-bake of existing agents is required.

Change-Id: If33d3ebe9028e7a8470b7218e115b21ac98808e0