5530 FIX Clicking on "Edit dashboard" accidentially copied dashboards with "publish to others"
When one clicks on "Edit dashboard" to edit the currently visible dashboard and the dashboard
is not owned by this user, the dashboard is cloned in the context of the current user. This
is intended and equal to, for example, the views.

What was unintended is the fact that the "public" flag of the cloned dashboards was not forced
to be forced. This way all users (with enough permissions) were publishing their cloned views
leading to strang situations where from time to time a user could see different dashboards.

The "Make this dashboard available for other users" is forced to be unchecked. This way a user
clones the dashboard for himself in the first place.

Change-Id: Id4e99498f697646460dcaf67091ff95a4260bd5b