5545 FIX Editing rules: Fixed adding date+user to rule comments
When editing a rule in WATO, there is the comment field available, which can e.g.
be used to write down information about the history of this rule. The icon on
the right of the text area is meant to add a new history entry to the beginning
of the comment. The way this text was added to existing comments was not really
working as intended.

It has now been changed to add a new line to the beginning, setting the cursor
on this field and position it at the end of the just added line.

Change-Id: Ic992f023b87be39512cc1afce155cc9874bad209