5558 FIX Rule list: Fixed visibility of rules in child folders for transitive folders
When viewing the list of rules in a ruleset via WATO, it shows the rules grouped by folder.
When the user views a transitive folder the rules defined in child folders were not listed.

It has now been changed to work like this: The list shows all rules of all parent and
all child folders. Folders of other, not involved, branches in the folder hierarchy
are not displayed.

Change-Id: I16da6d1875dcfe7d0275cae5cfb784e60f8ea1b9