5265 FIX cmk-update-agent: Enable Agent Updater to update itself on Windows
When triggering an update under Windows, the agent updater could not be replaced by its new downloaded version because it was still running from the corresponding file.

This situation is now mitigated by restarting the agent updater from within ../temp when running from plugins-dir.

This is relevant only for a manual execution of cmk-update-agent.exe

Change-Id: I190c83cd7520b716797c037d5788e114a0f449d3