Made counter for active Livestatus connections atomic.
We have seen Livestatus usage >100% out in the wild, which is total
nonsense.  This commit is trying to fix this, but there could be other
problems: We really, really have to introduce memory barriers and locks in
Livestatus, the current code is simply wrong in this respect.

This is a port of f253e7bd86f45e787b72164289ad2ea5055e4d67.

Change-Id: I153389f4a12ceaea596d8bb889da57207803db8d