5488 FIX veeam_jobs: Add support for new veeam API
The new veeam API only supports StartTimeLocal and StopTimeLocal. The old API
using StartTime and StopTime is not supported anymore. This lead to the error
that backup jobs using the new API were shown as "No complete backup". This
change adds support for StartTimeLocal and StopTimeLocal while maintaining
compatibility to the old API.

Change-Id: I787f79b8239a451232ce8ccb90d59a33d8d93f50