5660 FIX Windows agent: Removed unused nowin.exe
The <tt>nowin.exe</tt> was shipped as part of the windows agent for a long time.
This unused file was now reported as being infected with some kind of virus/trojan,
which was a false positive in all cases we had a look at. To get rid of these false
positives we have decided to drop this file now.

Change-Id: I1e76e845833c855bc9299cba034ffeca90a21bfc