5866 FIX Fixed performance issue on ruleset search and discovery page with a large number of rules
When rule matching was analyzed in WATO, which is done for example on the ruleset rule listing
page, on the analyze rulset page of hosts or on the service discovery page, a time consuming
match function needs to be called. This function has now been optimizied which should result in
a more performant processing of these pages with a larger number of rules.

Change-Id: Ib6dc8fcf9b26a573f6aa8aacb260d97322383070