5871 FIX Improved service view rendering performance when showing perfometers
When displaying service views which conain the "Perf-O-Meter" painter, the rendering
of the page, especially with a larger number of rows (>1000), could take a lot of time
(multiple seconds) for rendering.

This was caused by an unoptimized perfometer rendering/selection algorithm. A workaround
with older versions is to remove the Perf-O-Meter painter from the views.

Change-Id: Ia3273b52f79417ae0d67a1ac1974c4ca2dbe77f0