5875 FIX Git: Fixed error "git add --all ... is outside repository" with older git versions
When using older git versions (saw this with 1.7.1) an error could occur when enabling the
git integration of WATO. This was caused by an incompatibility with the "git add" command
used by WATO. It used absolute paths to add the new files that should be added to git instead
of relative paths to the current working directory.

This issue was present in all 1.4.0 releases.

Change-Id: Iaffb451bb6b16840926093fa4b21742956c73128