5834 FIX Fixed SNMPv3 credential handling in the event console.
Although it is not really useful, the list of engine IDs, for which a given
set of credential parameters should be used, can be empty.  Previously this
lead to a traceback in the EC log and missing credentials for SNMP traps.
This has been fixed.

Another bug was in the handling of the noAuthNoPriv security level, where
the wrong user name was used. This, too, has been fixed.

Change-Id: If3a83a2b34e1cb8f22d95de92618bb826809745f