check_mk_agent.linux, check_mk_agent.openwrt: specify message digest for encrypted agent output explicitly
For encrypted agent output the default message digest algorithm of OpenSSL was
used before. Prior to OpenSSL 1.1 the default is MD5 which is what the Check_MK
server is expecting as well. Starting from OpenSSL 1.1. the default message
digest algorithm changed to SHA-256 which leads to problems in the
communication between hosts with a new OpenSSL version and the Check_MK server.
Now the message digest algorithm is specified explicitly as MD5 to circumvent
any ambiguity.

Change-Id: Ib200efb11ac7937f1af9c2fb48072dd5701ddd07