agent_kubernetes: don't use a piggyback prefix for nodes
Kubernetes objects received by the agent_kubernetes are prefixed by
the type of object (e.g. pod_ for Pods). For Kubernetes nodes this
is unhandy since in setups which monitor a Kubernetes node via the
check_mk_agent and the agent_kubernetes this often requires an
additional piggyback translation rule.

Change-Id: I6af813423b7021bbbee02a19e8fe06f87c2f28b9