7774 FIX Fix installation issue on older debian based distros
The werk #7344 introduced compatibility issues with Linux setups that don't
support linux capabilities (for different reasons). One reason may be a kernel
that does not support the capabilities.

We have added a fallback to the Checkmk debian packages that tries to use the
linux capabilities (as described in #7344). Once that fails it falls back to
setting the SETUID bit on the binaries, just like Checkmk did it before.

Change-Id: I4c563f067ec638531d769f0c36a170b7ae84d397