7487 FIX mk_logwatch: Consider 'Do not deploy' in bakery
When setting up a logwatch rule in the agent bakery, you can choose 'Do not deploy the Logwatch plugin'.
This option has not been effective so far.

Now, when creating the logwatch sections, all configured sections in the hierarchy used, until a 'Do not deploy the Logwatch plugin' or 'Deploy the Logwatch plugin without configuration' rule is met, at which point all sections defined upwards in the folder hierarchy are ignored.

This also fixes FEED-3726

Change-Id: I3bf59655d5666a5f261104c891bba79184d33521