7423 FIX mk_postgres: fix syntax error
Version 1.5.0p17 and 1.5.0p18 of the mk_postgres plugin gave a
syntax error when parsing databases. This results in empty
output between [databases_start] and [databases_end] in the
plugin output. As a consequence the checks postgres_stats,
postgres_bloat, postgres_connections, postgres_locks and
postgres_query_duration become stale.

To resolve the error the fixed version of mk_postgres has to be
deployed to the affected hosts.

Change-Id: I8dc5e88e3a7584538f2ba09407c54a27a1837ca4