7711 New Windows Agent
This release introduces new Checkmk Windows Agent

We are currently working on a user manual chapter that describes the new agent in detail.
It will be finished during the beta phase of Checkmk 1.6.

Important Notes:
The Location of the New Agent is "C:\Program Files (x86)\check_mk_service"
The Name of Service is "checkmkservice"
The Log file is located in the "C:\Users\Public"
New Agent doesn't support Windows XP or Windows 2003
If you need to support such platforms, please use the "Legacy agent" which can be downloaded
from the Agent page in WATO.
User Data for New Agent, like plugins and configurations are now in the folder "C:\ProgramData\CheckMK\Agent"
Configuration file is check_mk.user.yml, located in the folder "C:\ProgramData\CheckMK\Agent"
New Agent can re-use configuration and plugins of the Legacy Windows Agent.

Change-Id: I1fe25d654be2e0174add498f372fba5b286b45e9