Quick fix for reducing "Testbuild" runtime
The optimizations of the Python compilation take a long time.  Since the
packages produced by the test builds are not used for other things than
testing the build multiple times a day we can skip the optimizations
during this build execution without disadvantages.

Removing this decreases the job execution time from 3 hours to 1 hour
and 15 minutes.

Change-Id: I8ddbd8dc3d066bf237be5e1807370f150b5b007c