7490 FIX mk_logwatch: enforce utf-8 encoding of config files
Previously it was merely assumed that the config files are utf-8 encoded.
To properly address issues with matching of non-ASCII characters we now
require that the config files are utf-8 encoded.

Configuration files that cannot be decoded will be skipped, and the corresponding
service will go to {{UNKNOWN}}, displaying an appropriate message.

Change-Id: I05579503159ff06013f1134327572597111c4400