7262 Initial version of new Business Intelligence visualization
This werk introduces the initial version of the new business intelligence visualiziation, which was mentioned at our last conference
https://www.youtube.com/watch?v=zYfIBJkbKxA

You can navigate to the new visualization through an icon, which is shown in the classical view.
The new visualization enables you to view aggregations in a different perspective.

A more elaborate documentation will be provided later on. Some first notes:

You can create layout templates for a group of aggregations. These templates can be assigned in the aggregation configuration.
You can also specifically assign a layout to one aggregation.

Change-Id: I40327cc5e356b13664a61b2d13e62d6c380aba47