Protect snapshot unpacking, cmk --notify and the mknotifyd with a lock
When a snapshot on a remote site is unpacked lock_checkmk_configuration
is now used to avoid further modifications of the configuration.
For the same reason the configuration reload of the mknotifyd and the
command line cmk --notify are now protected with the configuration lock.
The mknotifyd only tries to aquire the lock to remain operational if
the lock is not available immediately. In this case the mknotifyd will
use the old configuration for another cycle.

CMK-2089

Change-Id: Ib52c0bcc60fbe50a51cf3994075fae9ef6f8f47b